Added {% image %} tag.
Creates an <img> tag in HTML with width="x" and height="y" attributes.
If max_height and/or max_width arguments are given to the tag,
the image is shrunk if necessary and saved under a new name in _site.
Good for thumbnails or making sure images fit in site design.

E.g.

{% image foo.png max_width=50 %}

will create _site/foo_50x?.png if foo.png is wider than 50 pixels.