Write index.html if permalink ends with slash.
If a post/page has a permalink like "/foo/bar/",
it will be in the file "foo/bar/index.html".

Fixes issue 64.

This is different from the old version that assumed
that anything that didn't end with .html was a
directory, so for the file "/atom.xml" it would create
"atom.xml/index.html".