Added filter string_escape.
This filer escapes quote chars and backslash in a string
so it can be used in a Javascript string literal.

E.g. if the Liquid variable "something" is:

IT'S FUN\"

And you have the following in the template:

var x = "{{ something | string_escape }}"

It will be rendered as:

var x = "IT\'S FUN\\\""